requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
Pygments==2.9.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tabulate==0.8.9
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
